backend/requirements.txt
fastapi[standard]==0.115.0
pydantic==2.9.2
supabase==2.8.1
groq==0.11.0
python-dotenv==1.0.1
uvicorn==0.32.0
